Refactored TextToSpeechTextOnly2.tsx and CompletedStory2.tsx into TextToSpeechTextOnly.tsx and CompletedStory.tsx
