+++ StoryQuest/app/Components/TextToSpeechTextOnly.tsx
@@ -1,84 +1,155 @@
-import React, { useState, useEffect } from "react";
-import useTextToSpeech from "@/Components/useTextToSpeech";
+import React, { useEffect, useState, useRef } from "react";
+
 interface TextToSpeechCompletedStoryProps {
-    text: string;
-    onComplete?: () => void; // when text to speech is done
+  text: string;
+  onComplete?: () => void;
 }
 
-// Text to speech phrases component, mainly used for story phrases text to speech
-const TextToSpeechTextOnly: React.FC<TextToSpeechCompletedStoryProps> = ({ text, onComplete }) => {
-    const [utterance, setUtterance] = useState<SpeechSynthesisUtterance | null>(null);
-    const [selectedVoice, setSelectedVoice] = useState<SpeechSynthesisVoice | null>(null);
+const TextToSpeechTextOnly: React.FC<TextToSpeechCompletedStoryProps> = ({
+  text,
+  onComplete,
+}) => {
+  const [isReady, setIsReady] = useState(false);
+  const voicesLoadedRef = useRef(false);
+  const utteranceRef = useRef<SpeechSynthesisUtterance | null>(null);
+  const attemptCountRef = useRef(0);
+  const maxAttempts = 5;
 
-    // Select voice
-    function selectVoice() {
-        const voices = speechSynthesis.getVoices();
-
-        // List of voices
-        const voiceNames = [
-            "Google UK English Male",  // Chrome
-            "Microsoft David - English (United States)",  // Edge, Windows
-            "Daniel",  // macOS
-        ];
-
-        return voices.find(voice => voiceNames.includes(voice.name)) || null; // if voices in list are not available it just uses the default voice on the device.
+  // Initialize speech synthesis
+  useEffect(() => {
+    if (typeof window === 'undefined' || !window.speechSynthesis) {
+      console.warn("Speech synthesis not supported");
+      return;
     }
 
-    // Loads in voices
-    useEffect(() => {
-        if (typeof window !== "undefined" && window.speechSynthesis) {
-            const synth = window.speechSynthesis;
+    const handleVoicesChanged = () => {
+      if (voicesLoadedRef.current) return;
+      voicesLoadedRef.current = true;
+      setIsReady(true);
+    };
 
-            const handleVoicesChanged = () => {
-                const voice = selectVoice(); // Selects a voice
-                setSelectedVoice(voice); // Save voice
-            };
+    // Try getting voices immediately
+    const voices = window.speechSynthesis.getVoices();
+    if (voices && voices.length > 0) {
+      voicesLoadedRef.current = true;
+      setIsReady(true);
+    } else {
+      // Wait for voices to load
+      window.speechSynthesis.addEventListener('voiceschanged', handleVoicesChanged);
+    }
 
-            synth.addEventListener("voiceschanged", handleVoicesChanged); // Synth Event listener
-            handleVoicesChanged();
+    // For Safari, which might not trigger voiceschanged
+    const safariFallbackTimer = setTimeout(() => {
+      if (!voicesLoadedRef.current) {
+        console.warn("Voice loading timed out, trying to proceed anyway");
+        setIsReady(true);
+      }
+    }, 1000);
 
-            // Cleanup on component unmount and remove event listener
-            return () => {
-                synth.removeEventListener("voiceschanged", handleVoicesChanged);
-            };
+    return () => {
+      window.speechSynthesis.removeEventListener('voiceschanged', handleVoicesChanged);
+      clearTimeout(safariFallbackTimer);
+      if (utteranceRef.current) {
+        window.speechSynthesis.cancel();
+      }
+    };
+  }, []);
+
+  // Select best available voice for the platform
+  const selectVoice = () => {
+    const voices = window.speechSynthesis.getVoices();
+    if (!voices || voices.length === 0) return null;
+
+    // Try to find a voice by name or language
+    const preferredVoiceNames = [
+      "Google UK English Male",
+      "Microsoft David",
+      "Daniel",
+      "Alex", // Good for Safari allegedly 
+    ];
+
+    // Try by name first
+    for (const name of preferredVoiceNames) {
+      const voice = voices.find(v => v.name.includes(name));
+      if (voice) return voice;
+    }
+
+    // Then try to find any English voice
+    const englishVoice = voices.find(v => v.lang.includes('en'));
+    if (englishVoice) return englishVoice;
+
+    // Fall back to the first voice
+    return voices[0];
+  };
+
+  // Speak the text when ready
+  useEffect(() => {
+    if (!text || !isReady) return;
+
+    const speakText = () => {
+      if (typeof window === 'undefined' || !window.speechSynthesis) return;
+
+      try {
+        // Cancel any ongoing speech
+        window.speechSynthesis.cancel();
+
+        // Create utterance
+        const utterance = new SpeechSynthesisUtterance(text.replace(/_/g, " "));
+        utteranceRef.current = utterance;
+
+        // Select voice
+        const voice = selectVoice();
+        if (voice) {
+          utterance.voice = voice;
         }
-    }, []);
 
-    // Text to speech
-    useEffect(() => {
-        if (typeof window !== "undefined" && window.speechSynthesis && selectedVoice && text) {
-            const synth = window.speechSynthesis;
+        // Set properties
+        utterance.rate = 0.9;
+        utterance.pitch = 1.0;
+        utterance.volume = 1.0;
 
-                const u = new SpeechSynthesisUtterance(text.replace(/_/g, ' '));
+        // Set callbacks
+        utterance.onend = () => {
+          console.log("Speech ended successfully");
+          utteranceRef.current = null;
+          onComplete?.();
+        };
 
+        utterance.onerror = (e) => {
+          console.warn("Speech synthesis error:", e);
+          utteranceRef.current = null;
+          
+          // Retry on error for mobile browsers
+          if (attemptCountRef.current < maxAttempts) {
+            attemptCountRef.current++;
+            setTimeout(speakText, 250);
+          } else {
+            // After max attempts, just call onComplete
+            onComplete?.();
+          }
+        };
 
-                // Set the selected voice
-                u.voice = selectedVoice;
+        // Workaround for some mobile browsers
+        setTimeout(() => {
+          window.speechSynthesis.speak(utterance);
+        }, 100);
+      } catch (err) {
+        console.error("Speech synthesis exception:", err);
+        onComplete?.();
+      }
+    };
 
-                // Update speech rate to 0.9
-                u.rate = 0.9;
+    // Start speaking with a small delay to help with mobile browsers
+    const timer = setTimeout(speakText, 200);
+    return () => {
+      clearTimeout(timer);
+      if (typeof window !== 'undefined' && window.speechSynthesis) {
+        window.speechSynthesis.cancel();
+      }
+    };
+  }, [text, isReady, onComplete]);
 
-                // onend event listener
-                u.onend = () => {
-                        onComplete?.(); // Trigger the callback when TTS is done
-                };
-
-                // speech error
-                u.onerror = (event) => {
-                };
-
-                // Play speech
-                synth.speak(u);
-
-                // Cleanup on component unmount
-                return () => {
-                    synth.cancel();
-                    setUtterance(null)
-                };
-        }
-    }, [text, selectedVoice]);
-
-    return null;
+  return null;
 };
 
 export default TextToSpeechTextOnly;
